requirements.txt
einops
h5py
matplotlib
monai
natsort
nibabel
numpy
pandas
piq
Pillow
scikit-image
scipy
statsmodels
timm
torch
torchmetrics
torchvision
tqdm
